apply flutter design systems: font, colors, textfield
Former-commit-id: 7ad4817de4e1552a87be2b019debf48a62433a09

+++ glowvy_mobile/lib/models/product/review_meta.dart
@@ -5,7 +5,7 @@ part 'review_meta.g.dart';
 @JsonSerializable(fieldRename: FieldRename.snake)
 class ReviewMeta {
   double averageRating;
-  double reviewCount;
+  int reviewCount;
   double rankingScore;
 //come
   ReviewMeta(
